Add KbuildComposer, implement two composition steps
The implemented steps copy the Linux source to a temporary directory and
generate an incomplete .config file based on the enabled and disabled
features.

To keep track of disabled features, the return value of the sample
method of Sampler was changed to List<Map<String, Boolean>>. This is
necessary because disabled options need to be set to n explicitly
in the .config file. Otherwise make olddefconfig may automatically set
some of the values to y.