Fix feature model reading and .config generation
Torte doesn't add unconstrained features to generated feature models,
so as a workaround Vari-Joern now tries to determine which features
are missing.
Previously, options that were not set in the default .config file
were not set by Vari-Joern either. This is solved by appending
these options to the .config file.