Parallelize iterations
Vari-Joern now uses a producer-consumer pattern to parallelize its
iterations. Simply starting new threads for each sample and running
the composer-sample pipeline sequentially would not be efficient
because the composers and analyzers would be too synchronized. The
goal is to prepare the next variant while the previous one is being
analyzed, which this approach would not achieve.