Move PC extraction code into composer strategies
I've moved CombinedPresenceConditionMapper back into the Kbuild
composer directory because it wasn't as generalizable as I thought
it would be. In particular, systems that use the standard C/C++
preprocessor yield different presence conditions for different
compiler flags, which wasn't captured by
CombinedPresenceConditionMapper. On the other hand, a future PC
extraction implementation for Fiasco wouldn't have this issue, so
fixes for this wouldn't be generally applicable.