Fix Dockerfile patch check (#73)
Vari-Joern's torte experiment scripts contain a hack that patches
the Dockerfile for torte's kmax image. To avoid patching it
unnecessarily, the scripts run grep beforehand. The path passed to the
grep commands was wrong in a few cases, leading to an error message in
torte's output but no functional bugs.